made BLOCK_STALE_MEASURE a uint32 and corrected propagateStakeUpdate logic and commented parameters

+++ src/contracts/interfaces/IDelayedService.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.9;
 
 interface IDelayedService {
-    function BLOCK_STALE_MEASURE() external view returns(uint256);    
+    function BLOCK_STALE_MEASURE() external view returns(uint32);    
 }
